pyproject.toml
[project]
name = "inazuma"
version = "3.2.7"
description = "A gui wrapper over viu"
license = "UNLICENSE"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffpyplayer>=4.5.3",
    "kivymd",
    "viu-media[standard]",
    "yt-dlp-ejs>=0.3.2",
    "yt-dlp[curl-cffi,default]>=2025.12.8",
]

[tool.uv.sources]
kivymd = { git = "https://github.com/kivymd/KivyMD.git" }
viu-media = { git = "https://github.com/viu-media/viu.git" }

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
